+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/Auto1Basket3Sempels.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.opmodes.auto;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.hardware.rev.RevTouchSensor;
@@ -12,8 +12,8 @@ import com.qualcomm.robotcore.hardware.Servo;
 import org.firstinspires.ftc.robotcore.external.navigation.AngleUnit;
 import org.firstinspires.ftc.robotcore.external.navigation.YawPitchRollAngles;
 
-@Autonomous(name = "Auto2Specimen", group = "Robot")
-public class Auto2SpecimenMethods extends LinearOpMode {
+@Autonomous(name = ";;;;2", group = "Robot")
+public class Auto1Basket3Sempels extends LinearOpMode {
     private DcMotor rightFront;
     private DcMotor leftBack;
     private DcMotor rightBack;
@@ -57,7 +57,6 @@ public class Auto2SpecimenMethods extends LinearOpMode {
         leftBack.setDirection(DcMotorSimple.Direction.REVERSE);
         rightBack.setDirection(DcMotorSimple.Direction.FORWARD);
 
-        upMotor.setDirection(DcMotorSimple.Direction.REVERSE);
 
         leftFront.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         rightFront.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
@@ -74,42 +73,30 @@ public class Auto2SpecimenMethods extends LinearOpMode {
         waitForStart();
 
         /*
-        /проезд в перёд до апарата
-        /подём до верхнего отсека
+        /проезд вперёд  до корзины
+        /подём до верхней корзины
         /опускание и отъезд назад
-        /поворот и проезд до зоны наблюдения
-        /поворот и взять образец
-        /поднятие длифта поворот и проезд влево и выравнивание об борт
-        /проезд прерёд и действия 1-3 строчку
-        /проезд назад до борта и поехать вправо
+        /поворот и проезд назад выравнивание об борт
+        /проезд вперёд влево и с пробой назад
+        /так 3 раза
          */
-        driveStraight(0.5, 50);
-        liftUp(2430, 0.3);
-        driveStraight(0.5, 10);
-        liftUp(1930, 0.3);
+        driveStraight(0.5, 30);
+        liftUp(4200, 0.5);
         liftClaw.setPosition(0.65);
-        driveStraight(-0.5, 30);
+        driveStraight(-0.5, 70);
         liftZero(-0.5);
-        driveRotate(0.5, 110);
-        driveStraight(0.5, 140);
-        driveRotate(0.5, 70);
-        driveStraight(0.5, 40);
-        liftClaw.setPosition(0);
-        driveStraight(0.5, 60);
-        driveSide(0.5, 130);
-        driveRotate(0.5, 180);
-        driveStraight(-0.5, 40);
-        driveStraight(0.5, 70);
-        liftUp(2430, 0.3);
-        driveStraight(0.5, 10);
-        liftUp(1930, 0.3);
-        liftClaw.setPosition(0.65);
-        driveStraight(-0.5, 30);
-        liftZero(-0.5);
-        driveStraight(0.5, -70);
-        driveSide(0.5, 150);
-
-
+        driveRotate(0.5, 40);
+        driveRotate(0.5, 270);
+        driveStraight(0.5, 30);
+        driveStraight(-0.5, 90);
+        driveRotate(0.5, 40);
+        driveStraight(0.5, 100);
+        driveStraight(-0.5, 100);
+        driveRotate(0.5, 30);
+        driveStraight(0.5, 100);
+        driveStraight(-0.5, 100);
+        driveRotate(0.5, 30);
+        driveStraight(0.5, 100);
 
 
     }
@@ -128,7 +115,7 @@ public class Auto2SpecimenMethods extends LinearOpMode {
         rightBack.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
 
 
-        while (opModeIsActive() && Math.abs(leftFront.getCurrentPosition()) < PULSES_PER_CM * distance){
+        while (opModeIsActive() && Math.abs(leftFront.getCurrentPosition()) < PULSES_PER_CM * distance) {
 
             leftFront.setPower(speed);
             rightFront.setPower(speed);
@@ -169,8 +156,9 @@ public class Auto2SpecimenMethods extends LinearOpMode {
         rightBack.setPower(0);
         sleep(300);
     }
+
     // без минуса по часовой
-    public void driveRotate (double speed, double angle) {
+    public void driveRotate(double speed, double angle) {
 
         while (opModeIsActive() && Math.abs(getDegrees()) < angle) {
             imu.resetYaw();
@@ -194,18 +182,17 @@ public class Auto2SpecimenMethods extends LinearOpMode {
         upMotor.setMode(DcMotor.RunMode.RUN_TO_POSITION);
         upMotor.setPower(power);
 
-        while (opModeIsActive() && upMotor.isBusy()) {
+        //  while (opModeIsActive() && upMotor.isBusy()) {
 
-            telemetry.addData("Позиция для мотора", upMotor.getCurrentPosition());
-            telemetry.update();
-        }
-
-
+        //       telemetry.addData("Позиция для мотора", upMotor.getCurrentPosition());
+        //      telemetry.update();
+        //   }
 
 
     }
+
     public void liftZero(double power) {
-        while (opModeIsActive() && !limitSwitch.isPressed()){
+        while (opModeIsActive() && !limitSwitch.isPressed()) {
             upMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
             upMotor.setPower(power);
         }
@@ -220,5 +207,4 @@ public class Auto2SpecimenMethods extends LinearOpMode {
         return orientation.getYaw(AngleUnit.DEGREES);
     }
 }
-
