+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/auto/AutoSpecimen2.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.opmodes.auto;
 
 /*import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
