+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/test/TestTeleOpLift.java
@@ -1,11 +1,10 @@
-package org.firstinspires.ftc.teamcode;
+package org.firstinspires.ftc.teamcode.opmodes.test;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 
-import java.security.acl.LastOwnerException;
 @TeleOp (name = "TestTeleopLift")
 public class TestTeleOpLift extends LinearOpMode {
     DcMotor liftMotor;
