+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/opmodes/test/WW.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode.opmodes;
+package org.firstinspires.ftc.teamcode.opmodes.test;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
@@ -6,16 +6,17 @@ import com.qualcomm.robotcore.hardware.DcMotor;
 import com.qualcomm.robotcore.hardware.DcMotorSimple;
 import com.qualcomm.robotcore.hardware.Servo;
 
-@TeleOp(name = "extensionTest", group = "Robot")
-public class liftTest extends LinearOpMode {
+@TeleOp(name = "WW", group = "Robot")
+public class WW extends LinearOpMode {
     DcMotor rightFront;
     DcMotor leftBack;
     DcMotor rightBack;
     DcMotor leftFront;
+    DcMotor liftMotor;
     //DcMotor sss;
-    DcMotor extension;
+    DcMotor extensionMotor;
     // DcMotor sss;
-    DcMotor lift;
+    Servo liftClaw;
     Servo turnServo;
     Servo extensionClaw;
     double forward;
@@ -36,25 +37,26 @@ public class liftTest extends LinearOpMode {
 
 
     public void runOpMode() {
-        lift = hardwareMap.get(DcMotor.class, "lift");
+        liftMotor = hardwareMap.get(DcMotor.class, "lift");
         //sss = hardwareMap.get(DcMotor.class, "sss");
-        extension = hardwareMap.get(DcMotor.class, "extension");
+        extensionMotor = hardwareMap.get(DcMotor.class, "extension");
         extensionClaw = hardwareMap.get(Servo.class, "extensionClaw");
         turnServo = hardwareMap.get(Servo.class, "turnServo");
-        lift.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        lift.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        lift.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        extension.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
-        extension.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
-        extension.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        liftMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        liftMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        liftMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
+        extensionMotor.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
+        extensionMotor.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+        extensionMotor.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
         //sss.setMode(DcMotor.RunMode.STOP_AND_RESET_ENCODER);
         //sss.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
         //sss.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
 
-        //72
+        telemetry.update();
+//72
 
-        //extensionClaw = hardwareMap.get(Servo.class, "extensionClaw");
-        //extensionClaw.setPosition(0);
+        extensionClaw = hardwareMap.get(Servo.class, "extensionClaw");
+        extensionClaw.setPosition(0);
 
         leftFront = hardwareMap.get(DcMotor.class, "leftFront");
         rightFront = hardwareMap.get(DcMotor.class, "rightFront");
@@ -65,8 +67,8 @@ public class liftTest extends LinearOpMode {
         rightFront.setDirection(DcMotorSimple.Direction.FORWARD);
         leftBack.setDirection(DcMotorSimple.Direction.REVERSE);
         rightBack.setDirection(DcMotorSimple.Direction.FORWARD);
-        lift.setDirection(DcMotorSimple.Direction.REVERSE);
-        extension.setDirection(DcMotorSimple.Direction.REVERSE);
+        liftMotor.setDirection(DcMotorSimple.Direction.REVERSE);
+        extensionMotor.setDirection(DcMotorSimple.Direction.REVERSE);
         //sss.setDirection(DcMotorSimple.Direction.REVERSE);
 
         waitForStart();
@@ -80,13 +82,14 @@ public class liftTest extends LinearOpMode {
 
 
         while (opModeIsActive()) {
+            if (gamepad2.a && !a) {
+                num += 0.01;
+                extensionClaw.setPosition(num);
 
-            lift.setPower(-0.025);
 
-
-
-
-            telemetry.addData("Клешня",lift.getCurrentPosition());
+            }
+            a = gamepad2.a;
+            telemetry.addData("Клешня",extensionClaw.getPosition());
             telemetry.update();
         }
     }
